block delete of player of another account

+++ controllers/account_controller.go
@@ -1,6 +1,7 @@
 package controllers
 
 import (
+	"fmt"
 	"net/http"
 	"strconv"
 
@@ -121,7 +122,7 @@ func RegisterPlayer(context *gin.Context) {
 		return
 	}
 
-	if _, err := repository.GetPlayer(request.Name); err == nil {
+	if _, err := repository.GetPlayerByName(request.Name); err == nil {
 		context.JSON(http.StatusBadRequest, gin.H{"error": "Player name " + request.Name + " already exists!"})
 		context.Abort()
 		return
@@ -181,9 +182,27 @@ func ListPlayers(context *gin.Context) {
 	context.JSON(http.StatusOK, response)
 }
 
+// @tags     Account/Login
+// @summary  Delete a player
+// @Security ApiKeyAuth
+// @param    player_id path     int true "Id of the player to delete"
+// @success  204
+// @router   /account/player/{player_id} [delete]
 func DeletePlayer(context *gin.Context) {
+	claims := auth.GetTokenClaims(context)
+	accountId := claims.Id
 	playerId := context.Param("playerId")
 	playerIdAsInt, _ := strconv.Atoi(playerId)
+
+	playersInAccount, _ := repository.GetPlayerById(strconv.Itoa(accountId))
+	fmt.Println(accountId)
+	fmt.Println(playersInAccount.Id)
+	fmt.Println(playersInAccount.Name)
+	if playersInAccount.Account_id != accountId {
+		context.JSON(http.StatusUnauthorized, gin.H{"error": "You can't delete a player that is not associated with your account."})
+		context.Abort()
+		return
+	}
 
 	if err := repository.DeletePlayer(playerIdAsInt); err != nil {
 		context.JSON(http.StatusInternalServerError, gin.H{"error": err.Error()})
